+++ lib/src/gesture/extended_image_slide_page_handler.dart
@@ -1,3 +1,4 @@
+import 'package:extended_image/src/extended_image_typedef.dart';
 import 'package:flutter/material.dart';
 import '../extended_image_utils.dart';
 import 'extended_image_gesture_utils.dart';
@@ -9,9 +10,16 @@ import 'extended_image_slide_page.dart';
 
 /// for loading/failed widget
 class ExtendedImageSlidePageHandler extends StatefulWidget {
-  const ExtendedImageSlidePageHandler(this.child, this.extendedImageSlidePageState);
+  const ExtendedImageSlidePageHandler({
+    @required this.child,
+    this.extendedImageSlidePageState,
+    this.heroBuilderForSlidingPage,
+  });
   final Widget child;
   final ExtendedImageSlidePageState extendedImageSlidePageState;
+
+  ///build Hero only for sliding page
+  final HeroBuilderForSlidingPage heroBuilderForSlidingPage;
   @override
   ExtendedImageSlidePageHandlerState createState() =>
       ExtendedImageSlidePageHandlerState();
@@ -20,6 +28,21 @@ class ExtendedImageSlidePageHandler extends StatefulWidget {
 class ExtendedImageSlidePageHandlerState
     extends State<ExtendedImageSlidePageHandler> {
   Offset _startingOffset;
+  ExtendedImageSlidePageState _extendedImageSlidePageState;
+  @override
+  void didChangeDependencies() {
+    _extendedImageSlidePageState = widget.extendedImageSlidePageState ??
+        context.findAncestorStateOfType<ExtendedImageSlidePageState>();
+    super.didChangeDependencies();
+  }
+
+  @override
+  void didUpdateWidget(covariant ExtendedImageSlidePageHandler oldWidget) {
+    _extendedImageSlidePageState = widget.extendedImageSlidePageState ??
+        context.findAncestorStateOfType<ExtendedImageSlidePageState>();
+    super.didUpdateWidget(oldWidget);
+  }
+
   @override
   Widget build(BuildContext context) {
     Widget result = GestureDetector(
@@ -29,15 +52,15 @@ class ExtendedImageSlidePageHandlerState
       child: widget.child,
       behavior: HitTestBehavior.translucent,
     );
-
-    if (widget.extendedImageSlidePageState != null &&
-        widget.extendedImageSlidePageState.widget.slideType ==
-            SlideType.onlyImage) {
-      final ExtendedImageSlidePageState extendedImageSlidePageState = widget.extendedImageSlidePageState;
+    if (_extendedImageSlidePageState != null) {
+      result = widget.heroBuilderForSlidingPage?.call(result) ?? result;
+    }
+    if (_extendedImageSlidePageState != null &&
+        _extendedImageSlidePageState.widget.slideType == SlideType.onlyImage) {
       result = Transform.translate(
-        offset: extendedImageSlidePageState.offset,
+        offset: _extendedImageSlidePageState.offset,
         child: Transform.scale(
-          scale: extendedImageSlidePageState.scale,
+          scale: _extendedImageSlidePageState.scale,
           child: result,
         ),
       );
@@ -52,14 +75,14 @@ class ExtendedImageSlidePageHandlerState
   Offset _updateSlidePagePreOffset;
   void _handleScaleUpdate(ScaleUpdateDetails details) {
     ///whether gesture page
-    if (widget.extendedImageSlidePageState != null && details.scale == 1.0) {
+    if (_extendedImageSlidePageState != null && details.scale == 1.0) {
       //var offsetDelta = (details.focalPoint - _startingOffset);
 
       final double delta = (details.focalPoint - _startingOffset).distance;
 
       if (doubleCompare(delta, minGesturePageDelta) > 0) {
         _updateSlidePagePreOffset ??= details.focalPoint;
-        widget.extendedImageSlidePageState.slide(
+        _extendedImageSlidePageState.slide(
             details.focalPoint - _updateSlidePagePreOffset,
             extendedImageSlidePageHandlerState: this);
         _updateSlidePagePreOffset = details.focalPoint;
@@ -68,10 +91,10 @@ class ExtendedImageSlidePageHandlerState
   }
 
   void _handleScaleEnd(ScaleEndDetails details) {
-    if (widget.extendedImageSlidePageState != null &&
-        widget.extendedImageSlidePageState.isSliding) {
+    if (_extendedImageSlidePageState != null &&
+        _extendedImageSlidePageState.isSliding) {
       _updateSlidePagePreOffset = null;
-      widget.extendedImageSlidePageState.endSlide(details);
+      _extendedImageSlidePageState.endSlide(details);
       return;
     }
   }